addressed the review comments
